backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
sqlalchemy
alembic
python-dotenv
openai
requests
jinja2
PyJWT
prometheus-client
sentry-sdk
google-generativeai
bcrypt
email-validator
